+++ data.py
@@ -1,9 +1,3 @@
-BASE = 'https://stellarburgers.nomoreparties.site'
-LOGIN_PAGE = 'https://stellarburgers.nomoreparties.site/login'
-REGISTER = 'https://stellarburgers.nomoreparties.site/register'
-FORGOT_PASSWORD = 'https://stellarburgers.nomoreparties.site/forgot-password'
-ACCOUNT_PROFILE = 'https://stellarburgers.nomoreparties.site/account/profile'
-ACCOUNT = 'https://stellarburgers.nomoreparties.site/account'
 LOGIN = 'vlad_tsygankov_111@mail.ru'
 PASSWORD = '123456'
 BAD_PASSWORD = '123'
