+++ url.py
@@ -1,0 +1,7 @@
+class URL:
+    BASE = 'https://stellarburgers.nomoreparties.site'
+    LOGIN_PAGE = 'https://stellarburgers.nomoreparties.site/login'
+    REGISTER = 'https://stellarburgers.nomoreparties.site/register'
+    FORGOT_PASSWORD = 'https://stellarburgers.nomoreparties.site/forgot-password'
+    ACCOUNT_PROFILE = 'https://stellarburgers.nomoreparties.site/account/profile'
+    ACCOUNT = 'https://stellarburgers.nomoreparties.site/account'
